+++ src/main.cpp
@@ -3432,6 +3432,7 @@ void YourClassName::sendNetworkSpectrumFrame(const std::vector<float> &frequenci
     frame["sampleRate"] = pendingSettings.sampleRate;
     frame["bandwidth"] = pendingSettings.bandwidth;
     frame["modulationType"] = pendingSettings.modulationType;
+    frame["inputMode"] = pendingSettings.inputMode;
     frame["fftLength"] = targetCount;
     frame["sourceFftLength"] = dataCount;
     frame["minFrequency"] = frameMinFrequency;